\name{addTwo-package}
\alias{addTwo-package}
\alias{addTwo}
\docType{package}
\title{
addTwo
}
\description{
A fuction used for returning the input adding 2.
}
\details{
\tabular{ll}{
Package: \tab addTwo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-18\cr
License: \tab GPL version 2 or later\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Yu Ouyang

Maintainer: Yu Ouyang <418614yo@student.eur.nl>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

# input 3 into the function, and the result will be 5.
a <- addTwo(3)

}
